+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/playlist/DefaultHlsPlaylistTracker.java
@@ -853,6 +853,14 @@ public final class DefaultHlsPlaylistTracker
             playlistSnapshot != oldPlaylist
                 ? playlistSnapshot.targetDurationUs
                 : (playlistSnapshot.targetDurationUs / 2);
+      } else if (playlistSnapshot == oldPlaylist) {
+        // To prevent infinite requests when the server responds with CAN-BLOCK-RELOAD=YES but does
+        // not actually block until the playlist updates, wait for half the target duration before
+        // retrying.
+        durationUntilNextLoadUs =
+            playlistSnapshot.partTargetDurationUs != C.TIME_UNSET
+                ? playlistSnapshot.partTargetDurationUs / 2
+                : playlistSnapshot.targetDurationUs / 2;
       }
       earliestNextLoadTimeMs =
           currentTimeMs + Util.usToMs(durationUntilNextLoadUs) - loadEventInfo.loadDurationMs;